Dev

Closes #30, #23, #6, and #7

See merge request sae-semestre4-groupexx/api-server!41

+++ app/Http/Controllers/Api/CommentaireController.php
@@ -8,6 +8,8 @@ use App\Http\Resources\CommentaireResource;
 use App\Models\Commentaire;
 use DateTime;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use OpenApi\Attributes as OA;
 
 class CommentaireController extends Controller
 {
@@ -21,11 +23,51 @@ class CommentaireController extends Controller
 
     }
 
+    #[OA\Post(
+        path: "/api/commentaire",
+        operationId: "create",
+        description: "Ajouter un commentaire dans la base de données",
+        requestBody: new OA\RequestBody(
+            required: true,
+        ),
+        tags: ["Commentaires"],
+        responses: [
+            new OA\Response(
+                response: 200,
+                description: "Création d'un commentaire",
+                content: new OA\JsonContent(properties: [
+                    new OA\Property(property: "status", type: "boolean"),
+                    new OA\Property(property: "message", type: "string"),
+                    new OA\Property(property: "commentaire")
+                ],type: "object")
+            ),
+            new OA\Response(
+                response: 422,
+                description: "Erreur",
+                content: new OA\JsonContent(properties: [
+                    new OA\Property(property: "message", type: "string"),
+                    new OA\Property(
+                        property: "errors",
+                        properties: [
+                            new OA\Property(property: "commentaire|date_com|note|etat", type: "array", items: new OA\Items(type: "string"))
+                        ],
+                        type: "object",
+                    )
+                ],type: "object")
+            ),
+        ]
+    )]
     /**
      * Store a newly created resource in storage.
      */
     public function store(CommentaireRequest $request)
     {
+        if (Gate::denies('store-commentaire')) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Vous n\'êtes pas autorisé à ajouter un commentaire !'
+            ], 403);
+        }
         $commentaire = new Commentaire();
         $commentaire->commentaire = $request->commentaire;
         $commentaire->date_com = new dateTime();
@@ -60,12 +102,44 @@ class CommentaireController extends Controller
         //
     }
 
+    #[OA\Put(
+        path: "/commentaires/{id}",
+        operationId: "update",
+        description: "Modifier un commentaire dans la base",
+        requestBody: new OA\RequestBody(
+            required: true,
+        ), tags: ["Commentaires"],
+        parameters: [
+            new OA\Parameter(
+                name: "id",
+                description: "Identifiant du commentaire",
+                in: "path", required: "true",
+                schema: new OA\Schema(type: "integer"))
+        ],
+        responses: [
+            new OA\Response(
+                response: 200,
+                description: "Modification d'un commentaire",
+                content: new OA\JsonContent(properties: [
+                    new OA\Property(property: "status", type: "boolean"),
+                    new OA\Property(property: "message", type: "string"),
+                    new OA\Property(property: "commentaire")
+                ], type: "object")
+            ),
+        ],
+    )]
     /**
      * Update the specified resource in storage.
      */
     public function update(CommentaireRequest $request, int $id)
     {
         $commentaire = Commentaire::findOrFail($id);
+        if (Gate::denies('update-commentaire', $commentaire)) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Vous n\'êtes pas autorisé à modifier ce commentaire !'
+            ], 403);
+        }
         $commentaire->update($request->all());
 
         $commentaire->commentaire = $request->commentaire;
@@ -95,6 +169,40 @@ class CommentaireController extends Controller
 
     }
 
+    #[OA\Delete(
+        path: "/commentaires/{id}",
+        operationId: "destroy",
+        description: "Supprime un commentaire",
+        tags: ["Commentaires"],
+        parameters: [
+            new OA\Parameter(
+                name: "id",
+                description: "Identifiant du commentaire",
+                in: "path", required: "true",
+                schema: new OA\Schema(type: "integer"))
+        ],
+        responses: [
+            new OA\Response(
+                response: 200,
+                description: "Supprime un commentaire",
+                content: new OA\JsonContent(properties: [
+                    new OA\Property(property: "status", type: "boolean"),
+                    new OA\Property(property: "message", type: "string"),
+                ], type: "object")
+            ),
+            new OA\Response(
+                response: 404,
+                description: "Commentaire non trouvée",
+                content: new OA\JsonContent(properties: [
+                    new OA\Property(property: "message", type: "string"),
+                    new OA\Property(property: "errors", properties: [
+                        new OA\Property(property: "id", type: "array", items: new OA\Items(type: "string"))
+                    ], type: "object"
+                    ),
+                ], type: "object")
+            )
+        ]
+    )]
     /**
      * Remove the specified resource from storage.
      */
@@ -102,6 +210,12 @@ class CommentaireController extends Controller
     {
         try {
             $commentaire = Commentaire::findOrFail($id);
+            if (Gate::denies('delete-commentaire', $commentaire)) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'Vous n\'êtes pas autorisé à supprimer ce commentaire !'
+                ], 403);
+            }
             $commentaire->delete();
             return response()->json([
                 'status' => 'success',
